+++ src/letsbuilda/pypi/sync_client.py
@@ -4,7 +4,7 @@ from http import HTTPStatus
 from typing import Final, Self
 
 import xmltodict
-from requests import Session
+from httpx import Client
 
 from .exceptions import PackageNotFoundError
 from .models import JSONPackageMetadata, Package, RSSPackageMetadata
@@ -16,12 +16,12 @@ class PyPIServices:
     NEWEST_PACKAGES_FEED_URL: Final[str] = "https://pypi.org/rss/packages.xml"
     PACKAGE_UPDATES_FEED_URL: Final[str] = "https://pypi.org/rss/updates.xml"
 
-    def __init__(self: Self, http_session: Session) -> None:
-        self.http_session = http_session
+    def __init__(self: Self, http_client: Client) -> None:
+        self.http_client = http_client
 
     def get_rss_feed(self: Self, feed_url: str) -> list[RSSPackageMetadata]:
         """Get the new packages RSS feed."""
-        response_text = self.http_session.get(feed_url).text
+        response_text = self.http_client.get(feed_url).text
         rss_data = xmltodict.parse(response_text)["rss"]["channel"]["item"]
         return [RSSPackageMetadata.build_from(package_data) for package_data in rss_data]
 
@@ -35,7 +35,7 @@ class PyPIServices:
             url = f"https://pypi.org/pypi/{package_title}/{package_version}/json"
         else:
             url = f"https://pypi.org/pypi/{package_title}/json"
-        response = self.http_session.get(url)
+        response = self.http_client.get(url)
         if response.status_code == HTTPStatus.NOT_FOUND:
             raise PackageNotFoundError(package_title, package_version)
         return JSONPackageMetadata.from_dict(response.json())